SWC-6923 - Refactor to create CannedRejectionDialog
CannedRejectionDialog will be used in the existing RejectDataAccessRequestModal and in a new component used to reject profile validation requests
